+++ __tests__/bundle/bundle-anchor/snapshot.js
@@ -28,9 +28,6 @@ paths:
                 foo: TEST
 components: {}
 
-No configurations were defined in extends -- using built in recommended configuration by default.
-Warning! This default behavior is going to be deprecated soon.
-
 Woohoo! Your OpenAPI definitions are valid. 🎉
 
 bundling ./test.yaml...